Issue 40831: Overflowing integer values on PK for targetedms.TransitionChromInfo (#220)

+++ src/org/labkey/targetedms/TargetedMSController.java
@@ -1263,7 +1263,7 @@ public class TargetedMSController extends SpringActionController
         @Override
         public ModelAndView getView(ChromatogramForm form, BindException errors)
         {
-            int precursorId = form.getId();
+            long precursorId = form.getId();
             Precursor precursor = PrecursorManager.getPrecursor(getContainer(), precursorId, getUser());
             if (precursor == null)
             {
@@ -1359,7 +1359,7 @@ public class TargetedMSController extends SpringActionController
         @Override
         public ModelAndView getView(ChromatogramForm form, BindException errors)
         {
-            int precursorId = form.getId();
+            long precursorId = form.getId();
             MoleculePrecursor precursor = MoleculePrecursorManager.getPrecursor(getContainer(), precursorId, getUser());
             if (precursor == null)
             {
@@ -1441,7 +1441,7 @@ public class TargetedMSController extends SpringActionController
         @Override
         public ModelAndView getView(ChromatogramForm form, BindException errors)
         {
-            int peptideId = form.getId();
+            long peptideId = form.getId();
             Peptide peptide = PeptideManager.getPeptide(getContainer(), peptideId);
             if (peptide == null)
             {
@@ -1758,7 +1758,7 @@ public class TargetedMSController extends SpringActionController
 
     public static class ChromatogramForm extends AbstractChartForm
     {
-        private int _id;
+        private long _id;
         private boolean _syncY = false;
         private boolean _syncX = false;
         private boolean _splitGraph = false;
@@ -1772,12 +1772,12 @@ public class TargetedMSController extends SpringActionController
             setDefaultChartWidth(400);
         }
 
-        public int getId()
+        public long getId()
         {
             return _id;
         }
 
-        public void setId(int id)
+        public void setId(long id)
         {
             _id = id;
         }
@@ -1921,7 +1921,7 @@ public class TargetedMSController extends SpringActionController
         @Override
         public ModelAndView getView(ChromatogramForm form, BindException errors)
         {
-            int peptideId = form.getId();  // peptide Id
+            long peptideId = form.getId();  // peptide Id
 
             Peptide peptide = PeptideManager.getPeptide(getContainer(), peptideId);
             if(peptide == null)
@@ -2021,7 +2021,7 @@ public class TargetedMSController extends SpringActionController
         @Override
         public ModelAndView getView(ChromatogramForm form, BindException errors)
         {
-            int moleculeId = form.getId();
+            long moleculeId = form.getId();
 
             Molecule molecule = MoleculeManager.getMolecule(getContainer(), moleculeId);
             if (molecule == null)
@@ -4081,12 +4081,12 @@ public class TargetedMSController extends SpringActionController
         private List<ReplicateAnnotation> _replicateAnnotationValueList;
 
         // fields for proteomics
-        private int _peptideId;
-        private int _precursorId;
+        private long _peptideId;
+        private long _precursorId;
         private List<Peptide> _peptideList;
 
         // fields for small molecule
-        private int _moleculeId;
+        private long _moleculeId;
         private int _moleculePrecursorId;
         private List<Molecule> _moleculeList;
 
@@ -4130,32 +4130,32 @@ public class TargetedMSController extends SpringActionController
             _peptideGroupId = peptideGroupId;
         }
 
-        public int getPeptideId()
+        public long getPeptideId()
         {
             return _peptideId;
         }
 
-        public void setPeptideId(int peptideId)
+        public void setPeptideId(long peptideId)
         {
             _peptideId = peptideId;
         }
 
-        public int getPrecursorId()
+        public long getPrecursorId()
         {
             return _precursorId;
         }
 
-        public void setPrecursorId(int precursorId)
+        public void setPrecursorId(long precursorId)
         {
             _precursorId = precursorId;
         }
 
-        public int getMoleculeId()
+        public long getMoleculeId()
         {
             return _moleculeId;
         }
 
-        public void setMoleculeId(int moleculeId)
+        public void setMoleculeId(long moleculeId)
         {
             _moleculeId = moleculeId;
         }
